backend/requirements.txt
aiogram==2.14.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
environs==14.1.0
frozenlist==1.5.0
idna==3.10
marshmallow==3.25.1
multidict==6.1.0
packaging==24.2
pillow==11.1.0
propcache==0.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
